requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pandas==2.2.2
numpy==1.26.4
httpx==0.27.2
jinja2==3.1.4
